Split the file to ease testing and review
