+++ components/tools/OmeroJava/test/integration/LightAdminRolesChownTest.java
@@ -1,0 +1,249 @@
+/*
+ *------------------------------------------------------------------------------
+ *  Copyright (C) 2017 University of Dundee & Open Microscopy Environment.
+ *  All rights reserved.
+ *
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along
+ *  with this program; if not, write to the Free Software Foundation, Inc.,
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ *------------------------------------------------------------------------------
+ */
+package integration;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import omero.gateway.util.Requests;
+import omero.model.Dataset;
+import omero.model.DatasetImageLink;
+import omero.model.ExperimenterGroup;
+import omero.model.ExperimenterI;
+import omero.model.IObject;
+import omero.model.Image;
+import omero.model.OriginalFile;
+import omero.model.Project;
+import omero.model.ProjectDatasetLink;
+import omero.model.enums.AdminPrivilegeChown;
+import omero.model.enums.AdminPrivilegeSudo;
+import omero.sys.EventContext;
+
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+
+/**
+ *
+ * @author Jean-Marie Burel &nbsp;&nbsp;&nbsp;&nbsp;
+ * <a href="mailto:j.burel@dundee.ac.uk">j.burel@dundee.ac.uk</a>
+ * @since 5.4
+ */
+public class LightAdminRolesChownTest extends RolesTests {
+
+    /**
+     * Test that light admin can, having the <tt>Chown</tt> privilege,
+     * transfer the data between two users (normalUser and anotherUser).
+     * Test also that light admin, if sudoed, cannot transfer ownership,
+     * because light admin sudoes as a non-admin non-group-owner user.
+     * In case of private group the transfer of an Image severs the link between the Dataset and Image.
+     * For this unlinking, only the Chown permissions are sufficient, no other permissions are necessary.
+     * <tt>Chown</tt> privilege is sufficient also
+     * to transfer ownership of annotations (tag and file attachment are tested here),
+     * but just in case of private and read-only groups, which is in line with the
+     * general behavior of the <tt>Chown</tt> command.
+     * @param isSudoing if to test a success of workflows where Sudoed in
+     * @param permChown if to test a user who has the <tt>Chown</tt> privilege
+     * @param groupPermissions to test the effect of group permission level
+     * @throws Exception unexpected
+     * @see <a href="https://downloads.openmicroscopy.org/resources/experimental/tests/graph-permissions/0.1/testChown.pptx">graphical explanation</a>
+     */
+    @Test(dataProvider = "isSudoing and Chown privileges cases")
+    public void testChown(boolean isSudoing, boolean permChown, String groupPermissions)
+            throws Exception {
+        /* Define the conditions for the chown of the image is passing.*/
+        final boolean chownImagePassing = permChown && !isSudoing;
+        /* Chown of the annotations on the image is passing when
+         * chownImagePassing is true in higher permissions groups (read-annotate and read-write)
+         * only.*/
+        final boolean annotationsChownExpectSuccess = chownImagePassing &&
+                (groupPermissions.equals("rw----") || groupPermissions.equals("rwr---"));
+
+        final EventContext normalUser = newUserAndGroup(groupPermissions);
+        final EventContext anotherUser = newUserAndGroup(groupPermissions);
+        /* Create a Dataset as the normalUser and import into it */
+        loginUser(normalUser);
+        Dataset dat = mmFactory.simpleDataset();
+        Dataset sentDat = (Dataset) iUpdate.saveAndReturnObject(dat);
+        List<IObject> originalFileAndImage = importImageWithOriginalFile(sentDat);
+        OriginalFile originalFile = (OriginalFile) originalFileAndImage.get(0);
+        Image image = (Image) originalFileAndImage.get(1);
+
+        /* Annotate the imported image with Tag and file attachment.*/
+        List<IObject> annotOriginalFileAnnotationTagAndLinks = annotateImageWithTagAndFile(image);
+
+        /* Set up the basic permissions for this test.*/
+        List<String> permissions = new ArrayList<String>();
+        permissions.add(AdminPrivilegeSudo.value);
+        if (permChown) permissions.add(AdminPrivilegeChown.value);
+        final EventContext lightAdmin;
+        lightAdmin = loginNewAdmin(true, permissions);
+        sudo(new ExperimenterI(normalUser.userId, false));
+
+        /* Take care of workflows which do not use sudo.*/
+        if (!isSudoing) {
+            loginUser(lightAdmin);
+        }
+        /* Check that the value of canChown boolean matches chownPassingWhenNotSudoing
+         * boolean in each case.*/
+        Assert.assertEquals(getCurrentPermissions(image).canChown(), chownImagePassing);
+        /* Get into correct group context and check all cases.*/
+        client.getImplicitContext().put("omero.group", Long.toString(normalUser.groupId));
+        /* lightAdmin tries to chown the image.*/
+        doChange(client, factory, Requests.chown().target(image).toUser(anotherUser.userId).build(), chownImagePassing);
+        /* ChecK the results of the chown when lightAdmin is sudoed,
+         * which should fail in any case.*/
+        if (isSudoing) {
+            assertOwnedBy(image, normalUser);
+            assertOwnedBy(originalFile, normalUser);
+            assertOwnedBy(annotOriginalFileAnnotationTagAndLinks, normalUser);
+        /* Check the chown was successful for both the image and the annotations
+         * when the permissions for chowning both
+         * the image as well as the annotations on it are sufficient.*/
+        } else if (chownImagePassing && annotationsChownExpectSuccess) {
+            assertOwnedBy(image, anotherUser);
+            assertOwnedBy(originalFile, anotherUser);
+            /* Annotations will be chowned because
+             * groupPermissions are private or read-only (captured in boolean
+             * annotationsChownExpectSuccess).*/
+            assertOwnedBy(annotOriginalFileAnnotationTagAndLinks, anotherUser);
+        /* Check the chown was successful for the image but not the annotations
+         * in case the annotationsChownExpectSuccess is false, i.e. in read-only and private group.*/
+        } else if (chownImagePassing && !annotationsChownExpectSuccess){
+            assertOwnedBy(image, anotherUser);
+            assertOwnedBy(originalFile, anotherUser);
+            assertOwnedBy(annotOriginalFileAnnotationTagAndLinks, normalUser);
+        } else {
+        /* In the remaining case, the chown will fail, as the chownPassingWhenNotSudoing
+         * is false because permChown was not given. All objects belong to normalUser.*/
+            assertOwnedBy(image, normalUser);
+            assertOwnedBy(originalFile, normalUser);
+            assertOwnedBy(annotOriginalFileAnnotationTagAndLinks, normalUser);
+        }
+        /* In any case, the image must be in the right group.*/
+        assertInGroup(image, normalUser.groupId);
+    }
+
+
+    /**
+     * Light admin (lightAdmin) tries to transfer the ownership of all the data of a user (normalUser)
+     * to another user. The data are in 2 groups, of which the original data owner (normalUser)
+     * is member, the recipient of the data is member of just one of the groups. Chown privilege
+     * is sufficient for lightAdmin to perform the workflow.
+     * @param isPrivileged if to test a user who has the <tt>Chown</tt> privilege
+     * @param groupPermissions to test the effect of group permission level
+     * @throws Exception unexpected
+     * @see <a href="https://downloads.openmicroscopy.org/resources/experimental/tests/graph-permissions/0.1/testChownAllBelongingToUserLightAdmin.pptx">graphical explanation</a>
+     */
+    @Test(dataProvider = "isPrivileged cases")
+    public void testChownAllBelongingToUser(boolean isPrivileged, String groupPermissions) throws Exception {
+        /* Chown privilege is sufficient for the workflow.*/
+        final boolean chownPassing = isPrivileged;
+        final EventContext normalUser = newUserAndGroup(groupPermissions);
+        ExperimenterGroup anotherGroup = newGroupAddUser(groupPermissions, normalUser.userId, false);
+        final EventContext recipient = newUserInGroup(anotherGroup, false);
+        /* Set up the light admin's permissions for this test.*/
+        List<String> permissions = new ArrayList<String>();
+        if (isPrivileged) permissions.add(AdminPrivilegeChown.value);
+        final EventContext lightAdmin;
+        lightAdmin = loginNewAdmin(true, permissions);
+        /* normalUser creates two sets of Project/Dataset/Image hierarchy in their default group.*/
+        loginUser(normalUser);
+        client.getImplicitContext().put("omero.group", Long.toString(normalUser.groupId));
+        Image image1 = mmFactory.createImage();
+        Image image2 = mmFactory.createImage();
+        Image sentImage1 = (Image) iUpdate.saveAndReturnObject(image1);
+        Image sentImage2 = (Image) iUpdate.saveAndReturnObject(image2);
+        Dataset dat1 = mmFactory.simpleDataset();
+        Dataset dat2 = mmFactory.simpleDataset();
+        Dataset sentDat1 = (Dataset) iUpdate.saveAndReturnObject(dat1);
+        Dataset sentDat2 = (Dataset) iUpdate.saveAndReturnObject(dat2);
+        Project proj1 = mmFactory.simpleProject();
+        Project proj2 = mmFactory.simpleProject();
+        Project sentProj1 = (Project) iUpdate.saveAndReturnObject(proj1);
+        Project sentProj2 = (Project) iUpdate.saveAndReturnObject(proj2);
+        DatasetImageLink linkOfDatasetImage1 = linkParentToChild(sentDat1, sentImage1);
+        DatasetImageLink linkOfDatasetImage2 = linkParentToChild(sentDat2, sentImage2);
+        ProjectDatasetLink linkOfProjectDataset1 = linkParentToChild(sentProj1, sentDat1);
+        ProjectDatasetLink linkOfProjectDataset2 = linkParentToChild(sentProj2, sentDat2);
+
+        /* normalUser creates two sets of Project/Dataset?Image hierarchy in the other group (anotherGroup).*/
+        client.getImplicitContext().put("omero.group", Long.toString(anotherGroup.getId().getValue()));
+        Image image1AnotherGroup = mmFactory.createImage();
+        Image image2AnotherGroup = mmFactory.createImage();
+        Image sentImage1AnootherGroup = (Image) iUpdate.saveAndReturnObject(image1AnotherGroup);
+        Image sentImage2AnotherGroup = (Image) iUpdate.saveAndReturnObject(image2AnotherGroup);
+        Dataset dat1AnotherGroup = mmFactory.simpleDataset();
+        Dataset dat2AnotherGroup = mmFactory.simpleDataset();
+        Dataset sentDat1AnotherGroup = (Dataset) iUpdate.saveAndReturnObject(dat1AnotherGroup);
+        Dataset sentDat2AnotherGroup = (Dataset) iUpdate.saveAndReturnObject(dat2AnotherGroup);
+        Project proj1AnotherGroup = mmFactory.simpleProject();
+        Project proj2AnotherGroup = mmFactory.simpleProject();
+        Project sentProj1AnootherGroup = (Project) iUpdate.saveAndReturnObject(proj1AnotherGroup);
+        Project sentProj2AnotherGroup = (Project) iUpdate.saveAndReturnObject(proj2AnotherGroup);
+        DatasetImageLink linkOfDatasetImage1AnotherGroup = linkParentToChild(sentDat1AnotherGroup, sentImage1AnootherGroup);
+        DatasetImageLink linkOfDatasetImage2AnotherGroup = linkParentToChild(sentDat2AnotherGroup, sentImage2AnotherGroup);
+        ProjectDatasetLink linkOfProjectDataset1AnotherGroup = linkParentToChild(sentProj1AnootherGroup, sentDat1AnotherGroup);
+        ProjectDatasetLink linkOfProjectDataset2AnotherGroup = linkParentToChild(sentProj2AnotherGroup, sentDat2AnotherGroup);
+        /* lightAdmin tries to transfers all normalUser's data to recipient.*/
+        loginUser(lightAdmin);
+        /* In order to be able to operate in both groups, get all groups context.*/
+        mergeIntoContext(client.getImplicitContext(), ALL_GROUPS_CONTEXT);
+        /* Check on one selected object only (sentProj1AnotherGroup) the value
+         * of canChown. The value must match the chownPassing boolean.*/
+        Assert.assertEquals(getCurrentPermissions(sentProj1AnootherGroup).canChown(), chownPassing);
+        /* Check that transfer proceeds only if chownPassing boolean is true.*/
+        doChange(client, factory, Requests.chown().targetUsers(normalUser.userId).toUser(recipient.userId).build(), chownPassing);
+        if (!chownPassing) {
+            /* Finish the test if no transfer of data could proceed.*/
+            return;
+        }
+        /* Check the transfer of all the data in normalUser's group was successful,
+         * first checking ownership of the first hierarchy set.*/
+        assertOwnedBy(sentProj1, recipient);
+        assertOwnedBy(sentDat1, recipient);
+        assertOwnedBy(sentImage1, recipient);
+        assertOwnedBy(linkOfDatasetImage1, recipient);
+        assertOwnedBy(linkOfProjectDataset1, recipient);
+        /* Check ownership of the second hierarchy set.*/
+        assertOwnedBy(sentProj2, recipient);
+        assertOwnedBy(sentDat2, recipient);
+        assertOwnedBy(sentImage2, recipient);
+        assertOwnedBy(linkOfDatasetImage2, recipient);
+        assertOwnedBy(linkOfProjectDataset2, recipient);
+        /* Check ownership of the objects in anotherGroup,
+         * first checking ownership of the first hierarchy.*/
+        assertOwnedBy(sentProj1AnootherGroup, recipient);
+        assertOwnedBy(sentDat1AnotherGroup, recipient);
+        assertOwnedBy(sentImage1AnootherGroup, recipient);
+        assertOwnedBy(linkOfDatasetImage1AnotherGroup, recipient);
+        assertOwnedBy(linkOfProjectDataset1AnotherGroup, recipient);
+        /* Check ownership of the second hierarchy set in anotherGroup.*/
+        assertOwnedBy(sentProj2AnotherGroup, recipient);
+        assertOwnedBy(sentDat2AnotherGroup, recipient);
+        assertOwnedBy(sentImage1AnootherGroup, recipient);
+        assertOwnedBy(linkOfDatasetImage2AnotherGroup, recipient);
+        assertOwnedBy(linkOfProjectDataset2AnotherGroup, recipient);
+    }
+
+
+}
